Runs a full DADA2 workflow.

This performs the following steps:

Learning the error rate for each run separately

Inferring sequence variants for each run and sample

Merging of feature tables across runs

Consensus chimera removal

Taxa assignment with Naive Bayes

Species assignment by exact alignment

Diagnostic plots of the error rates

You will usually want to preprocess the read files first with preprocess. Depending on your config the sequences might be represented by an MD5 hash. In this case the taxa table has an additional `sequence` column containing the real sequences.

denoise(object, config)

Arguments

object
An experiment data table as returned by find_read_files or a worflow object.

config
A configuration file as returned by config_denoise.

Value

A list containing the workflow results:

feature_table

Matrix of sequence variant abundances. Samples are and sequences are columns.

taxonomy

Matrix of taxonomy assignments. Rows are sequences and columns are taxonomy ranks.

errors

The error profiles estimated by DADA2. One for each run and read direction (forward/reverse).

error_plots

Plotted error profiles. One for each run and read direction (forward/reverse).

passed_reads

How many reads were kept in each step. Rows are samples and columns are workflow steps.

classified

The proportion of sequence variants that could be where a specific taxa rank could be classified.